pyproject.toml
[project]
name = "bot-brain"
version = "0.1.0"
description = ""
authors = [
    {name = "aLMASIK",email = "suleimenovalmas75@gmail.com"}
]
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.131.0,<0.132.0)",
    "uvicorn (>=0.41.0,<0.42.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "black (>=26.1.0,<27.0.0)",
    "groq (>=1.0.0,<2.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
